distributed_auth_system/token_auth_service/requirements.txt
flask
werkzeug
requests
pyjwt